initiator/requirements.txt
pulsar-client==2.7.2